requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.44.0
django-on-heroku==1.1.2
django-phonenumber-field==5.1.0
django-storages==1.11.1
django-tinymce==3.3.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
phonenumbers==8.12.21
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.7
sendgrid==6.7.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
